+++ src/FluentlyHttpClient/GraphQL/GqlRequest.cs
@@ -1,10 +1,17 @@
-﻿namespace FluentlyHttpClient.GraphQL
+﻿using System;
+
+namespace FluentlyHttpClient.GraphQL
 {
 	/// <summary>
 	/// Request object for GraphQL requests.
 	/// </summary>
-	public class GqlQuery // deprecated: rename to GqlRequest or GqlRequestPayload?
+	public class GqlRequest
 	{
+		/// <summary>
+		/// Gets or sets the operation name.
+		/// </summary>
+		public string? OperationName { get; set; }
+
 		/// <summary>
 		/// Gets or sets GraphQL query.
 		/// </summary>
@@ -14,5 +21,10 @@
 		/// Gets or sets GraphQL query variables.
 		/// </summary>
 		public object? Variables { get; set; }
+	}
+
+	[Obsolete("Use 'GqlRequest' instead.")]
+	public class GqlQuery : GqlRequest
+	{
 	}
 }